Move test from ErrorHandlingTaskTest to WorkerSinkTaskTest
